pyproject.toml
[project]
name = "mcp-ai-toolkit"
version = "1.0.0"
description = "A comprehensive suite of Model Context Protocol (MCP) servers for AI-powered automation"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Sarjil"}
]
keywords = [
    "mcp",
    "model-context-protocol",
    "ai",
    "automation",
    "claude",
    "anthropic",
    "insurance",
    "research",
    "monitoring"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "anthropic>=0.51.0",
    "arxiv>=2.2.0",
    "mcp>=1.8.0",
    "nest-asyncio>=1.6.0",
    "python-dotenv>=1.1.0",
    "requests>=2.31.0",
]

[project.urls]
Homepage = "https://github.com/sarjil77/mcp-ai-toolkit"
Repository = "https://github.com/sarjil77/mcp-ai-toolkit"
Issues = "https://github.com/sarjil77/mcp-ai-toolkit/issues"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
